put all src files into a src folder for cleaner repo
